Add files via upload
The jupyter notebook is centered around the processing and analysis of textual data using Python. The notebook initiates by loading a dataset named emotion_sentiment_dataset.csv into a pandas DataFrame, demonstrating the initial steps required to handle real-world data in data science.

The notebook further integrates natural language processing (NLP) by utilizing the nltk library to manage and process English stopwords—common words that generally do not add much meaning to the text and are thus removed during the preprocessing phase. This step is crucial for refining the dataset for more effective analysis.

Moreover, the notebook details a structured text preprocessing pipeline that includes:

Converting all text to lowercase to maintain uniformity.
Tokenizing the text to break it down into individual words or meaningful tokens.
Removing stopwords to focus on significant words in the analysis.
Each function within the pipeline is methodically applied to the text data, demonstrating a comprehensive approach to preprocessing textual data within a DataFrame. The processed tokens are then appended as a new column in the DataFrame, showcasing the practical application of NLP techniques for preparing data for deeper analysis or modeling.