Add files via upload
This approach provides a practical example of web scraping and data handling, focuses on extracting and analyzing data from Reddit using Python and the PRAW (Python Reddit API Wrapper) library. The notebook begins with setting up PRAW, including authentication with Reddit's API, which allows the script to interact programmatically with Reddit.

The project specifically targets the "bitcoin" subreddit, using it as a case study to demonstrate data extraction techniques. The notebook outlines how to retrieve various pieces of information from posts in the subreddit, including titles, IDs, authors, creation timestamps, scores, upvote ratios, and URLs. This data is then used to populate a pandas DataFrame, which serves as the basis for further analysis.

The code samples within the notebook also demonstrate how to handle sets in Python—adding unique headlines to a set to avoid duplicates, and later converting this set into a DataFrame. The notebook is particularly focused on handling real-time data from a popular online platform, showcasing how Python can be used to gather and prepare data for analytical tasks.