Add files via upload
This jupyter notebook delves into the practical application of sentiment analysis using Python and various libraries such as Pandas, NLTK, and scikit-learn. The notebook begins by loading a dataset from a JSON file named 'Arts_Crafts_and_Sewing_5.json', which likely contains reviews or similar textual data. It showcases how to sample data from this dataset to get a glimpse of its structure and contents.

Further into the notebook, it introduces sentiment analysis techniques by incorporating the Bing Liu Lexicon from the NLTK library. This lexicon contains lists of positive and negative words used in opinion mining. The notebook demonstrates how to download and access these lists, and how to utilize them to assess the sentiment of text data. It shows the total number of words in the opinion lexicon and provides examples of both positive and negative words.

A significant portion of the notebook is dedicated to building a sentiment scoring function. This involves creating a dictionary where each word from the Bing Liu Lexicon is associated with a positive or negative score. Subsequently, the notebook defines a function bing_liu_score, which calculates the sentiment score for a given text based on the presence of words from this dictionary. This function tokenizes the input text, converts it to lowercase to ensure case insensitivity, and iterates through the words to sum up their respective sentiment scores.