Add files via upload
This jupyter notebook concentrates on sentiment analysis using Python. This notebook is structured around a typical sentiment analysis workflow which includes loading data, utilizing sentiment lexicons, and constructing a sentiment scoring system.

The notebook begins by loading a dataset named sentimentdataset.csv using the pandas library. This is likely a collection of text data for sentiment analysis, and the notebook provides a sample of this data to understand its structure and contents. Following data loading, the notebook introduces the Bing Liu Lexicon, a resource available through the NLTK library that includes lists of positive and negative opinion words. This is crucial for the sentiment analysis as it allows the identification and categorization of words based on their sentiment polarity.

In preparation for sentiment analysis, the notebook demonstrates how to create a dictionary that assigns a positive or negative score to each word from the Bing Liu Lexicon. This dictionary serves as the basis for the sentiment scoring function bing_liu_score, which calculates a sentiment score for a given piece of text. The function works by tokenizing the text, converting it to lowercase (to ensure uniformity), and then summing up the scores of all words that appear in the predefined dictionary.