Add files via upload
This jupyter notebook focuses on text summarization techniques using Python libraries such as NLTK, NumPy, and NetworkX. This educational notebook starts by introducing the basics of text summarization, and then it delves into practical implementation with an example.

The notebook begins by loading and processing a text file named "Text1.txt," which contains multiple sentences structured into a paragraph. It uses the Natural Language Toolkit (NLTK) to manipulate and prepare text data, including removing stop words and tokenizing sentences. This preprocessing step is crucial for simplifying the text and making the subsequent analysis more effective.

For the core of the text summarization task, the notebook demonstrates how to calculate sentence similarity using the cosine distance method. Each sentence in the paragraph is compared to every other to establish a similarity score. This involves converting sentences into vectors and measuring the cosine distance between these vectors to understand how sentences relate to each other semantically.

Once all sentences are compared, the notebook constructs a similarity matrix and converts this matrix into a graph using the NetworkX library. This graph represents sentences as nodes, with edges weighted by their similarity scores. The PageRank algorithm is then applied to this graph to rank the sentences based on their importance and connectivity to other sentences.

Finally, the notebook concludes by extracting and displaying the top-ranked sentences as the summarized content of the input text. This approach not only illustrates basic concepts of text summarization but also shows how to implement these concepts programmatically to develop a functional text summarizer.