Add files via upload
This jupyter notebook is an extension of text summarization tasks with added complexity of text translation. This notebook is sophisticated in its approach to handle text data by integrating translation tools, which makes it suitable for multilingual text analysis.

Initially, the notebook sets up by importing necessary libraries such as NLTK for text processing, NetworkX for graph-based analysis, NumPy for numerical operations, and Google's translation API through the googletrans library for translation tasks. The main function, process_and_translate, demonstrates the process of reading a text file, cleaning it, and translating the sentences into Kannada and Hindi. This showcases how to prepare textual data for cross-lingual applications, an essential skill in global data analysis and software applications.

Furthermore, the notebook employs a cosine similarity approach to compute similarities between sentences in the text. It creates a similarity matrix which is then converted into a graph structure using NetworkX. The notebook applies the PageRank algorithm to rank sentences based on their importance and connectivity to other sentences in the graph. This ranked list of sentences forms the basis of the text summarization output.

The final output includes both the top-ranked sentences for summarization and their translated versions in Kannada and Hindi, demonstrating a practical application of multilingual NLP tasks.