Add files via upload
This jupyter notebook explores the field of topic modeling on a dataset of U.N. General Debate speeches. The notebook is a structured guide on how to apply Natural Language Processing (NLP) techniques to extract themes or topics from large volumes of text.

Initially, the notebook begins by importing the necessary Python libraries and loading the dataset using pandas from a local CSV file named 'un-general-debates-blueprint.csv'. The dataset appears to consist of textual data from various speeches, from which the notebook showcases a couple of text excerpts to give an idea of the content.

The analysis progresses by splitting the speeches into paragraphs using regular expressions, which is a common preprocessing step in NLP to simplify and segment text data for better analysis. The notebook then focuses on extracting features from the text using the Term Frequency-Inverse Document Frequency (TF-IDF) vectorizer from the scikit-learn library. This vectorization approach helps in reducing the impact of frequently appearing words that might otherwise dominate the analysis in simpler bag-of-words models.

By applying TF-IDF, the notebook sets the groundwork for conducting topic modeling, which is likely to follow in the subsequent sections of the notebook. This process involves identifying and grouping similar words into topics that represent the main themes discussed in the speeches. Topic modeling is essential in understanding and summarizing large datasets by extracting the underlying topics that pervade the texts.