Add files via upload
This jupyter notebook focuses on analyzing a dataset containing spam text messages. The primary objective of this notebook is to read and preliminarily analyze the data to understand its structure and the type of information it contains.

The notebook starts by loading the dataset named "SPAM text message 20170820 - Data.csv" using pandas, a popular Python library for data manipulation and analysis. The initial step includes displaying the first few rows of the dataset with the head() function to give a snapshot of the content, which helps in getting a quick overview of the data's format and the types of messages included.

Following the initial data viewing, the notebook explores the dataset's structure more deeply. It includes commands to list the column names and data types, which are essential for understanding the kinds of variables involved and how they are stored. This is further supplemented by the info() method, providing a concise summary of the DataFrame. This summary includes details about column names, non-null values, and memory usage, which are crucial for assessing data quality and preparing for more detailed analysis.

Additionally, the notebook uses the describe() function to generate descriptive statistics that summarize the central tendency, dispersion, and shape of the dataset’s numeric distributions. Although the content primarily discusses textual data (i.e., spam messages), this function would be pertinent if there are numeric values or encoded categories within the data that can be quantitatively analyzed.